requirements.txt
# Core ML Dependencies
tensorflow==2.13.0
azureml-core==1.53.0
azureml-pipeline==1.53.0
azureml-pipeline-steps==1.53.0
azureml-train==1.53.0
azureml-train-automl==1.53.0
azureml-mlflow==1.53.0
mlflow==2.6.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.0.76

# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Monitoring & Logging
applicationinsights==0.11.10
opencensus==0.11.0
opencensus-ext-azure==1.1.8

# Security & Testing
bandit==1.7.5
safety==2.3.5
semgrep==1.40.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Infrastructure
kubernetes==27.2.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.17.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.5.2
tqdm==4.66.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.4.0 